Updates to simulation class and parameters.py
