+++ Model/parameters.py
@@ -1,10 +1,13 @@
 import numpy as np
 import igraph
+from backend import Base
 
 
 # used to hold all values necessary to the simulation based on the various modes
-class Simulation:
-    def __init__(self, templates_path, name, path, mode, separator):
+class Simulation(Base):
+    def __init__(self, templates_path, name, path, separator, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+
         self.name = name    # name of the simulation, used to name files
         self.path = path    # path to the main simulation directory
 
@@ -61,110 +64,48 @@ class Simulation:
 
         # the following instance variables are only necessary for a new simulation. these include arrays for storing
         # the cell values, graphs for cell neighbors, and other values that aren't included in the template files
-        if mode == 0:
-            # start the simulation at step 1
-            self.beginning_step = 1
 
-            # the spatial resolution of the space
-            self.spat_res = 0.00001  # 10 um
-            self.spat_res2 = self.spat_res ** 2
+        # start the simulation at step 1
+        self.beginning_step = 1
 
-            # the temporal resolution for the simulation
-            self.step_dt = 1800  # dt of each simulation step (1800 sec)
-            self.move_dt = 200  # dt for incremental movement (200 sec)
-            self.diffuse_dt = 0.5  # dt for stable diffusion model (0.5 sec)
+        # the spatial resolution of the space
+        self.spat_res = 0.00001  # 10 um
+        self.spat_res2 = self.spat_res ** 2
 
-            # min and max radius lengths are used to calculate linear growth of the radius over time in 2D
-            self.max_radius = 0.000005  # 5 um
-            self.min_radius = self.max_radius / 2 ** 0.5
-            self.pluri_growth = (self.max_radius - self.min_radius) / self.pluri_div_thresh
-            self.diff_growth = (self.max_radius - self.min_radius) / self.diff_div_thresh
+        # the temporal resolution for the simulation
+        self.step_dt = 1800  # dt of each simulation step (1800 sec)
+        self.move_dt = 200  # dt for incremental movement (200 sec)
+        self.diffuse_dt = 0.5  # dt for stable diffusion model (0.5 sec)
 
-            # holds all indices of cells that will divide or be removed during each step
-            self.cells_to_divide = np.array([], dtype=int)
-            self.cells_to_remove = np.array([], dtype=int)
+        # min and max radius lengths are used to calculate linear growth of the radius over time in 2D
+        self.max_radius = 0.000005  # 5 um
+        self.min_radius = self.max_radius / 2 ** 0.5
+        self.pluri_growth = (self.max_radius - self.min_radius) / self.pluri_div_thresh
+        self.diff_growth = (self.max_radius - self.min_radius) / self.diff_div_thresh
 
-            # much like the cell arrays add any graphs to this list for automatic addition/removal
-            self.graph_names = ["neighbor_graph", "jkr_graph"]
+        # holds all indices of cells that will divide or be removed during each step
+        self.cells_to_divide = np.array([], dtype=int)
+        self.cells_to_remove = np.array([], dtype=int)
 
-            # create graphs used to all cells and their neighbors, initialize them with the number of cells
-            self.neighbor_graph = igraph.Graph(self.number_cells)
-            self.jkr_graph = igraph.Graph(self.number_cells)
+        # much like the cell arrays add any graphs to this list for automatic addition/removal
+        self.graph_names = ["neighbor_graph", "jkr_graph"]
 
-            # the diffusion constant for the molecule gradients and the radius of search for diffusion points
-            self.diffuse = 0.00000000005    # 50 um^2/s
-            self.diffuse_radius = self.spat_res * 0.707106781187
+        # create graphs used to all cells and their neighbors, initialize them with the number of cells
+        self.neighbor_graph = igraph.Graph(self.number_cells)
+        self.jkr_graph = igraph.Graph(self.number_cells)
 
-            # much like the cell arrays add any gradient names to list this so that a diffusion function can
-            # act on them automatically
-            self.extracellular_names = ["fgf4_values", "fgf4_alt"]
+        # the diffusion constant for the molecule gradients and the radius of search for diffusion points
+        self.diffuse = 0.00000000005    # 50 um^2/s
+        self.diffuse_radius = self.spat_res * 0.707106781187
 
-            # calculate the size of the array holding the diffusion points and create gradient objects
-            self.gradient_size = np.ceil(self.size / self.spat_res).astype(int) + np.ones(3, dtype=int)
-            self.fgf4_values = np.zeros(self.gradient_size)
-            self.fgf4_alt = np.zeros(self.gradient_size)
+        # much like the cell arrays add any gradient names to list this so that a diffusion function can
+        # act on them automatically
+        self.extracellular_names = ["fgf4_values", "fgf4_alt"]
 
-            # used to hold the run times of functions
-            self.function_times = dict()
+        # calculate the size of the array holding the diffusion points and create gradient objects
+        self.gradient_size = np.ceil(self.size / self.spat_res).astype(int) + np.ones(3, dtype=int)
+        self.fgf4_values = np.zeros(self.gradient_size)
+        self.fgf4_alt = np.zeros(self.gradient_size)
 
-    def cell_types(self, *args):
-        """ go through the cell types adding them to the
-            simulation
-        """
-        self.holder = dict()
-        self.number_cells = 0
-        for cell_type in args:
-            begin = self.number_cells
-            self.number_cells += cell_type[1]
-            end = self.number_cells
-            self.holder[cell_type[0]] = (begin, end)
-
-    def cell_arrays(self, *args):
-        """ creates the Simulation instance arrays that
-            correspond to particular cell values
-        """
-        # go through all arguments passed
-        for array_params in args:
-            self.cell_array_names.append(array_params[0])
-
-            # get the length of the tuple
-            length = len(array_params)
-
-            # if the tuple passed is of length two, make a 1-dimensional array
-            if length == 2:
-                size = 0
-
-            # if the tuple
-            elif length == 3:
-                size = (0, array_params[2])
-
-            # raise error if otherwise
-            else:
-                raise Exception("tuples should have length 2 or 3")
-
-            # create an instance variable for the cell array with the specified size and type
-            self.__dict__[array_params[0]] = np.empty(size, dtype=array_params[1])
-
-    def initials(self, cell_type, array_name, func):
-        """ given a lambda function for the initial values
-            of a cell array this updates that accordingly
-        """
-        if cell_type == "all":
-            # get the cell array
-            cell_array = self.__dict__[array_name]
-
-            shape = list(cell_array.shape)
-            shape[0] = self.number_cells
-            array_type = cell_array.dtype
-            empty_array = np.empty(shape, dtype=array_type)
-            self.__dict__[array_name] = np.concatenate((cell_array, empty_array))
-
-            for i in range(self.number_cells):
-                self.__dict__[array_name][i] = func()
-
-        else:
-            begin = self.holder[cell_type][0]
-            end = self.holder[cell_type][1]
-
-            for i in range(begin, end):
-                self.__dict__[array_name][i] = func()
+        # used to hold the run times of functions
+        self.function_times = dict()
